refactor(data_samples): move sample datasets to /utils/data_samples
