+++ .playground/playgroud.tsx
@@ -12,7 +12,7 @@ import {
   Settings,
   mergeWithDefaultTheme,
 } from '../src';
-import { KIBANA_METRICS } from '../src/lib/series/utils/test_dataset_kibana';
+import { KIBANA_METRICS } from '../src/utils/data_samples/test_dataset_kibana';
 
 export class Playground extends React.Component {
   render() {
