+++ src/lib/series/domains/y_domain.test.ts
@@ -2,7 +2,7 @@ import { getGroupId, getSpecId, GroupId } from '../../../utils/ids';
 import { ScaleType } from '../../utils/scales/scales';
 import { RawDataSeries } from '../series';
 import { BasicSeriesSpec, DomainRange } from '../specs';
-import { BARCHART_1Y0G } from '../utils/test_dataset';
+import { BARCHART_1Y0G } from '../../../utils/data_samples/test_dataset';
 import {
   coerceYScaleTypes,
   getDataSeriesOnGroup,
