+++ src/lib/series/series.test.ts
@@ -12,7 +12,7 @@ import {
 } from './series';
 import { BasicSeriesSpec } from './specs';
 import { formatStackedDataSeriesValues } from './stacked_series_utils';
-import * as TestDataset from './utils/test_dataset';
+import * as TestDataset from '../../utils/data_samples/test_dataset';
 
 describe('Series', () => {
   test('Can split dataset into 1Y0G series', () => {
