+++ src/state/utils.test.ts
@@ -10,7 +10,7 @@ import {
   HistogramModeAlignments,
   LineSeriesSpec,
 } from '../lib/series/specs';
-import { BARCHART_1Y0G, BARCHART_1Y1G } from '../lib/series/utils/test_dataset';
+import { BARCHART_1Y0G, BARCHART_1Y1G } from '../utils/data_samples/test_dataset';
 import { LIGHT_THEME } from '../lib/themes/light_theme';
 import { AxisId, getGroupId, getSpecId, SpecId } from '../utils/ids';
 import { ScaleContinuous } from '../lib/utils/scales/scale_continuous';
