--- src/utils/data_generators/data_examples.ts
@@ -1,812 +1,0 @@
-export const GITHUB_DATASET = [
-  {
-    authorAssociation: 'Team Member',
-    vizType: 'Data Table',
-    issueType: 'Bug',
-    count: 14,
-  },
-  {
-    authorAssociation: 'Team Member',
-    vizType: 'Data Table',
-    issueType: 'Other',
-    count: 22,
-  },
-  {
-    authorAssociation: 'Team Member',
-    vizType: 'Heatmap',
-    issueType: 'Bug',
-    count: 12,
-  },
-  {
-    authorAssociation: 'Team Member',
-    vizType: 'Heatmap',
-    issueType: 'Other',
-    count: 6,
-  },
-  {
-    authorAssociation: 'Team Member',
-    vizType: 'Markdown',
-    issueType: 'Bug',
-    count: 6,
-  },
-  {
-    authorAssociation: 'Team Member',
-    vizType: 'Markdown',
-    issueType: 'Other',
-    count: 11,
-  },
-  {
-    authorAssociation: 'Team Member',
-    vizType: 'MetricVis',
-    issueType: 'Bug',
-    count: 16,
-  },
-  {
-    authorAssociation: 'Team Member',
-    vizType: 'MetricVis',
-    issueType: 'Other',
-    count: 8,
-  },
-  {
-    authorAssociation: 'Team Member',
-    vizType: 'Pie Chart',
-    issueType: 'Bug',
-    count: 7,
-  },
-  {
-    authorAssociation: 'Team Member',
-    vizType: 'Pie Chart',
-    issueType: 'Other',
-    count: 4,
-  },
-  {
-    authorAssociation: 'Team Member',
-    vizType: 'Tagcloud',
-    issueType: 'Bug',
-    count: 19,
-  },
-  {
-    authorAssociation: 'Team Member',
-    vizType: 'Tagcloud',
-    issueType: 'Other',
-    count: 13,
-  },
-  {
-    authorAssociation: 'Team Member',
-    vizType: 'TSVB',
-    issueType: 'Bug',
-    count: 86,
-  },
-  {
-    authorAssociation: 'Team Member',
-    vizType: 'TSVB',
-    issueType: 'Other',
-    count: 123,
-  },
-  {
-    authorAssociation: 'Team Member',
-    vizType: 'Timelion',
-    issueType: 'Bug',
-    count: 58,
-  },
-  {
-    authorAssociation: 'Team Member',
-    vizType: 'Timelion',
-    issueType: 'Other',
-    count: 93,
-  },
-  {
-    authorAssociation: 'Team Member',
-    vizType: 'Vega vis',
-    issueType: 'Bug',
-    count: 11,
-  },
-  {
-    authorAssociation: 'Team Member',
-    vizType: 'Vega vis',
-    issueType: 'Other',
-    count: 38,
-  },
-  {
-    authorAssociation: 'Team Member',
-    vizType: 'Point Series',
-    issueType: 'Bug',
-    count: 1,
-  },
-  {
-    authorAssociation: 'Team Member',
-    vizType: 'Point Series',
-    issueType: 'Other',
-    count: 1,
-  },
-  {
-    authorAssociation: 'Team Member',
-    vizType: 'Inspector',
-    issueType: 'Bug',
-    count: 15,
-  },
-  {
-    authorAssociation: 'Team Member',
-    vizType: 'Inspector',
-    issueType: 'Other',
-    count: 11,
-  },
-  {
-    authorAssociation: 'Community',
-    vizType: 'Data Table',
-    issueType: 'Bug',
-    count: 6,
-  },
-  {
-    authorAssociation: 'Community',
-    vizType: 'Data Table',
-    issueType: 'Other',
-    count: 24,
-  },
-  {
-    authorAssociation: 'Community',
-    vizType: 'Heatmap',
-    issueType: 'Bug',
-    count: 11,
-  },
-  {
-    authorAssociation: 'Community',
-    vizType: 'Heatmap',
-    issueType: 'Other',
-    count: 5,
-  },
-  {
-    authorAssociation: 'Community',
-    vizType: 'Markdown',
-    issueType: 'Bug',
-    count: 0,
-  },
-  {
-    authorAssociation: 'Community',
-    vizType: 'Markdown',
-    issueType: 'Other',
-    count: 1,
-  },
-  {
-    authorAssociation: 'Community',
-    vizType: 'MetricVis',
-    issueType: 'Bug',
-    count: 6,
-  },
-  {
-    authorAssociation: 'Community',
-    vizType: 'MetricVis',
-    issueType: 'Other',
-    count: 10,
-  },
-  {
-    authorAssociation: 'Community',
-    vizType: 'Pie Chart',
-    issueType: 'Bug',
-    count: 3,
-  },
-  {
-    authorAssociation: 'Community',
-    vizType: 'Pie Chart',
-    issueType: 'Other',
-    count: 5,
-  },
-  {
-    authorAssociation: 'Community',
-    vizType: 'Tagcloud',
-    issueType: 'Bug',
-    count: 2,
-  },
-  {
-    authorAssociation: 'Community',
-    vizType: 'Tagcloud',
-    issueType: 'Other',
-    count: 1,
-  },
-  {
-    authorAssociation: 'Community',
-    vizType: 'TSVB',
-    issueType: 'Bug',
-    count: 28,
-  },
-  {
-    authorAssociation: 'Community',
-    vizType: 'TSVB',
-    issueType: 'Other',
-    count: 51,
-  },
-  {
-    authorAssociation: 'Community',
-    vizType: 'Timelion',
-    issueType: 'Bug',
-    count: 29,
-  },
-  {
-    authorAssociation: 'Community',
-    vizType: 'Timelion',
-    issueType: 'Other',
-    count: 43,
-  },
-  {
-    authorAssociation: 'Community',
-    vizType: 'Vega vis',
-    issueType: 'Bug',
-    count: 2,
-  },
-  {
-    authorAssociation: 'Community',
-    vizType: 'Vega vis',
-    issueType: 'Other',
-    count: 9,
-  },
-  {
-    authorAssociation: 'Community',
-    vizType: 'Point Series',
-    issueType: 'Bug',
-    count: 2,
-  },
-  {
-    authorAssociation: 'Community',
-    vizType: 'Point Series',
-    issueType: 'Other',
-    count: 3,
-  },
-  {
-    authorAssociation: 'Community',
-    vizType: 'Inspector',
-    issueType: 'Bug',
-    count: 5,
-  },
-  {
-    authorAssociation: 'Community',
-    vizType: 'Inspector',
-    issueType: 'Other',
-    count: 8,
-  },
-];
-
-export const GROUPED_BAR_CHART = [
-  {
-    timestamp: '2018-08-13',
-    status: 'success',
-    count: 3942,
-  },
-  {
-    timestamp: '2018-08-13',
-    status: 'info',
-    count: 3502,
-  },
-  {
-    timestamp: '2018-08-13',
-    status: 'security',
-    count: 937,
-  },
-  {
-    timestamp: '2018-08-13',
-    status: 'warning',
-    count: 476,
-  },
-  {
-    timestamp: '2018-08-13',
-    status: 'error',
-    count: 248,
-  },
-  {
-    timestamp: '2018-08-14',
-    status: 'success',
-    count: 3939,
-  },
-  {
-    timestamp: '2018-08-14',
-    status: 'info',
-    count: 3506,
-  },
-  {
-    timestamp: '2018-08-14',
-    status: 'security',
-    count: 923,
-  },
-  {
-    timestamp: '2018-08-14',
-    status: 'warning',
-    count: 453,
-  },
-  {
-    timestamp: '2018-08-14',
-    status: 'error',
-    count: 275,
-  },
-  {
-    timestamp: '2018-08-15',
-    status: 'success',
-    count: 3888,
-  },
-  {
-    timestamp: '2018-08-15',
-    status: 'info',
-    count: 3539,
-  },
-  {
-    timestamp: '2018-08-15',
-    status: 'security',
-    count: 916,
-  },
-  {
-    timestamp: '2018-08-15',
-    status: 'warning',
-    count: 482,
-  },
-  {
-    timestamp: '2018-08-15',
-    status: 'error',
-    count: 295,
-  },
-  {
-    timestamp: '2018-08-16',
-    status: 'success',
-    count: 7,
-  },
-  {
-    timestamp: '2018-08-16',
-    status: 'info',
-    count: 6,
-  },
-  {
-    timestamp: '2018-08-16',
-    status: 'security',
-    count: 1,
-  },
-];
-
-export const MULTI_GROUPED_BAR_CHART = [
-  {
-    timestamp: '2018-08-13',
-    status: 'success',
-    os: 'win 7',
-    count: 817,
-  },
-  {
-    timestamp: '2018-08-13',
-    status: 'success',
-    os: 'ios',
-    count: 813,
-  },
-  {
-    timestamp: '2018-08-13',
-    status: 'success',
-    os: 'win xp',
-    count: 790,
-  },
-  {
-    timestamp: '2018-08-13',
-    status: 'success',
-    os: 'win 8',
-    count: 773,
-  },
-  {
-    timestamp: '2018-08-13',
-    status: 'success',
-    os: 'osx',
-    count: 749,
-  },
-  {
-    timestamp: '2018-08-13',
-    status: 'info',
-    os: 'win 7',
-    count: 715,
-  },
-  {
-    timestamp: '2018-08-13',
-    status: 'info',
-    os: 'ios',
-    count: 713,
-  },
-  {
-    timestamp: '2018-08-13',
-    status: 'info',
-    os: 'win xp',
-    count: 704,
-  },
-  {
-    timestamp: '2018-08-13',
-    status: 'info',
-    os: 'win 8',
-    count: 700,
-  },
-  {
-    timestamp: '2018-08-13',
-    status: 'info',
-    os: 'osx',
-    count: 670,
-  },
-  {
-    timestamp: '2018-08-13',
-    status: 'security',
-    os: 'win 7',
-    count: 201,
-  },
-  {
-    timestamp: '2018-08-13',
-    status: 'security',
-    os: 'ios',
-    count: 196,
-  },
-  {
-    timestamp: '2018-08-13',
-    status: 'security',
-    os: 'osx',
-    count: 195,
-  },
-  {
-    timestamp: '2018-08-13',
-    status: 'security',
-    os: 'win xp',
-    count: 176,
-  },
-  {
-    timestamp: '2018-08-13',
-    status: 'security',
-    os: 'win 8',
-    count: 169,
-  },
-  {
-    timestamp: '2018-08-13',
-    status: 'warning',
-    os: 'osx',
-    count: 103,
-  },
-  {
-    timestamp: '2018-08-13',
-    status: 'warning',
-    os: 'win 7',
-    count: 100,
-  },
-  {
-    timestamp: '2018-08-13',
-    status: 'warning',
-    os: 'win 8',
-    count: 95,
-  },
-  {
-    timestamp: '2018-08-13',
-    status: 'warning',
-    os: 'ios',
-    count: 90,
-  },
-  {
-    timestamp: '2018-08-13',
-    status: 'warning',
-    os: 'win xp',
-    count: 88,
-  },
-  {
-    timestamp: '2018-08-13',
-    status: 'error',
-    os: 'osx',
-    count: 60,
-  },
-  {
-    timestamp: '2018-08-13',
-    status: 'error',
-    os: 'ios',
-    count: 52,
-  },
-  {
-    timestamp: '2018-08-13',
-    status: 'error',
-    os: 'win xp',
-    count: 48,
-  },
-  {
-    timestamp: '2018-08-13',
-    status: 'error',
-    os: 'win 8',
-    count: 45,
-  },
-  {
-    timestamp: '2018-08-13',
-    status: 'error',
-    os: 'win 7',
-    count: 43,
-  },
-  {
-    timestamp: '2018-08-14',
-    status: 'success',
-    os: 'osx',
-    count: 811,
-  },
-  {
-    timestamp: '2018-08-14',
-    status: 'success',
-    os: 'ios',
-    count: 786,
-  },
-  {
-    timestamp: '2018-08-14',
-    status: 'success',
-    os: 'win 7',
-    count: 782,
-  },
-  {
-    timestamp: '2018-08-14',
-    status: 'success',
-    os: 'win xp',
-    count: 782,
-  },
-  {
-    timestamp: '2018-08-14',
-    status: 'success',
-    os: 'win 8',
-    count: 778,
-  },
-  {
-    timestamp: '2018-08-14',
-    status: 'info',
-    os: 'osx',
-    count: 718,
-  },
-  {
-    timestamp: '2018-08-14',
-    status: 'info',
-    os: 'win 8',
-    count: 717,
-  },
-  {
-    timestamp: '2018-08-14',
-    status: 'info',
-    os: 'win 7',
-    count: 701,
-  },
-  {
-    timestamp: '2018-08-14',
-    status: 'info',
-    os: 'ios',
-    count: 692,
-  },
-  {
-    timestamp: '2018-08-14',
-    status: 'info',
-    os: 'win xp',
-    count: 678,
-  },
-  {
-    timestamp: '2018-08-14',
-    status: 'security',
-    os: 'win xp',
-    count: 206,
-  },
-  {
-    timestamp: '2018-08-14',
-    status: 'security',
-    os: 'ios',
-    count: 197,
-  },
-  {
-    timestamp: '2018-08-14',
-    status: 'security',
-    os: 'win 7',
-    count: 180,
-  },
-  {
-    timestamp: '2018-08-14',
-    status: 'security',
-    os: 'osx',
-    count: 179,
-  },
-  {
-    timestamp: '2018-08-14',
-    status: 'security',
-    os: 'win 8',
-    count: 161,
-  },
-  {
-    timestamp: '2018-08-14',
-    status: 'warning',
-    os: 'win xp',
-    count: 97,
-  },
-  {
-    timestamp: '2018-08-14',
-    status: 'warning',
-    os: 'ios',
-    count: 93,
-  },
-  {
-    timestamp: '2018-08-14',
-    status: 'warning',
-    os: 'win 8',
-    count: 93,
-  },
-  {
-    timestamp: '2018-08-14',
-    status: 'warning',
-    os: 'osx',
-    count: 88,
-  },
-  {
-    timestamp: '2018-08-14',
-    status: 'warning',
-    os: 'win 7',
-    count: 82,
-  },
-  {
-    timestamp: '2018-08-14',
-    status: 'error',
-    os: 'win 7',
-    count: 62,
-  },
-  {
-    timestamp: '2018-08-14',
-    status: 'error',
-    os: 'ios',
-    count: 58,
-  },
-  {
-    timestamp: '2018-08-14',
-    status: 'error',
-    os: 'win xp',
-    count: 54,
-  },
-  {
-    timestamp: '2018-08-14',
-    status: 'error',
-    os: 'win 8',
-    count: 51,
-  },
-  {
-    timestamp: '2018-08-14',
-    status: 'error',
-    os: 'osx',
-    count: 50,
-  },
-  {
-    timestamp: '2018-08-15',
-    status: 'success',
-    os: 'osx',
-    count: 795,
-  },
-  {
-    timestamp: '2018-08-15',
-    status: 'success',
-    os: 'ios',
-    count: 781,
-  },
-  {
-    timestamp: '2018-08-15',
-    status: 'success',
-    os: 'win xp',
-    count: 780,
-  },
-  {
-    timestamp: '2018-08-15',
-    status: 'success',
-    os: 'win 7',
-    count: 766,
-  },
-  {
-    timestamp: '2018-08-15',
-    status: 'success',
-    os: 'win 8',
-    count: 766,
-  },
-  {
-    timestamp: '2018-08-15',
-    status: 'info',
-    os: 'win xp',
-    count: 723,
-  },
-  {
-    timestamp: '2018-08-15',
-    status: 'info',
-    os: 'ios',
-    count: 716,
-  },
-  {
-    timestamp: '2018-08-15',
-    status: 'info',
-    os: 'win 7',
-    count: 708,
-  },
-  {
-    timestamp: '2018-08-15',
-    status: 'info',
-    os: 'osx',
-    count: 701,
-  },
-  {
-    timestamp: '2018-08-15',
-    status: 'info',
-    os: 'win 8',
-    count: 691,
-  },
-  {
-    timestamp: '2018-08-15',
-    status: 'security',
-    os: 'osx',
-    count: 199,
-  },
-  {
-    timestamp: '2018-08-15',
-    status: 'security',
-    os: 'ios',
-    count: 188,
-  },
-  {
-    timestamp: '2018-08-15',
-    status: 'security',
-    os: 'win 8',
-    count: 188,
-  },
-  {
-    timestamp: '2018-08-15',
-    status: 'security',
-    os: 'win 7',
-    count: 179,
-  },
-  {
-    timestamp: '2018-08-15',
-    status: 'security',
-    os: 'win xp',
-    count: 162,
-  },
-  {
-    timestamp: '2018-08-15',
-    status: 'warning',
-    os: 'ios',
-    count: 103,
-  },
-  {
-    timestamp: '2018-08-15',
-    status: 'warning',
-    os: 'win 7',
-    count: 97,
-  },
-  {
-    timestamp: '2018-08-15',
-    status: 'warning',
-    os: 'win 8',
-    count: 97,
-  },
-  {
-    timestamp: '2018-08-15',
-    status: 'warning',
-    os: 'win xp',
-    count: 97,
-  },
-  {
-    timestamp: '2018-08-15',
-    status: 'warning',
-    os: 'osx',
-    count: 88,
-  },
-  {
-    timestamp: '2018-08-15',
-    status: 'error',
-    os: 'osx',
-    count: 64,
-  },
-  {
-    timestamp: '2018-08-15',
-    status: 'error',
-    os: 'win 7',
-    count: 60,
-  },
-  {
-    timestamp: '2018-08-15',
-    status: 'error',
-    os: 'ios',
-    count: 59,
-  },
-  {
-    timestamp: '2018-08-15',
-    status: 'error',
-    os: 'win 8',
-    count: 57,
-  },
-  {
-    timestamp: '2018-08-15',
-    status: 'error',
-    os: 'win xp',
-    count: 55,
-  },
-];
