+++ stories/bar_chart.tsx
@@ -28,9 +28,9 @@ import {
   timeFormatter,
   TooltipType,
 } from '../src/';
-import * as TestDatasets from '../src/lib/series/utils/test_dataset';
+import * as TestDatasets from '../src/utils/data_samples/test_dataset';
 
-import { KIBANA_METRICS } from '../src/lib/series/utils/test_dataset_kibana';
+import { KIBANA_METRICS } from '../src/utils/data_samples/test_dataset_kibana';
 
 import { TEST_DATASET_DISCOVER } from '../src/lib/series/utils/test_dataset_discover_per_30s';
